tg_bot/requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
httpx==0.27.2
requests>=2.31.0
pydantic==2.8.2
uvloop==0.20.0; sys_platform != 'win32'
openai>=1.51.0
elevenlabs>=1.7.0
replicate>=0.22.0
psycopg2-binary>=2.9.9
urllib3<2.0